session4/7_dockerised-app/requirements.txt
async-timeout==4.0.2
attrs==22.1.0
avro==1.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
Flask==2.2.2
Flask-Injector==0.14.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
inflection==0.5.1
injector==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
mysql==0.0.3
mysql-client==0.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
packaging==21.3
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dotenv==0.21.0
PyYAML==6.0
redis==4.3.5
requests==2.28.1
swagger-ui-bundle==0.0.9
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
